Update ij-display code to ImgLib2.
git-svn-id: http://code.imagej.net/svn/imagej/branches/imglib2@2703 be23795f-a60e-40d4-ba19-055798872e57

+++ ui/ui-swing/src/main/java/imagej/ui/swing/display/SwingImageDisplay.java
@@ -194,7 +194,7 @@ public class SwingImageDisplay implements AWTDisplay {
 	}
 
 	@Override
-	public int[] getCurrentPlanePosition() {
+	public long[] getCurrentPlanePosition() {
 		return controller.getPos();
 	}
